fix(cache): honor the store contract's signal, expiry, and test isolation
Address PR review feedback and the CI failures on #4461:

- Honor the optional AbortSignal: operations reject promptly at entry and
  between backend round-trips. Backend calls themselves are not cancelable,
  matching the registry's advisory-cancellation model.
- read() returns null for a record a backend retained past its deadline; the
  contract reserves null for missing or expired history.
- Move environment-dependent tests (pinning rollout flags, shared-backend
  predicates) to tests/integration/server/dependency-snapshot-store-wiring.test.ts
  with explicitly pinned env, keeping the unit files hermetic. This fixes the
  semantic-disposition lint and the env-sensitive coverage-shard assertions.
- Document why the 60s renewal-deadline slack is safe: shared history can
  expire at most a minute early, and the registry renews half a retention
  period before expiry.
- Format.
